stm32: remove last few mod sealed's.
